pyproject.toml
[tool.poetry]
name = "logica-api"
version = "0.1.0"
description = ""
authors = ["altergeist <altergeist.dpkg@gmail.com>"]
readme = "README.md"
packages = [{include = "logica_api"}]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
fastapi = "^0.101.1"
sqlalchemy = "^2.0.19"
black = "^23.7.0"
alembic = "^1.11.2"
pytz = "^2023.3"
pydantic-settings = "^2.0.3"
pandas = "^2.0.3"
python-multipart = "^0.0.6"
asyncpg = "^0.28.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
start = "logica_api.main:start"